@@ -1,5 +1,9 @@
 import {AfterViewInit, Component, OnInit, ViewChild} from '@angular/core';
-import {MatPaginator, MatSlideToggleChange, MatSnackBar, MatSort, MatTableDataSource} from '@angular/material';
+import { MatPaginator } from '@angular/material/paginator';
+import { MatSlideToggleChange } from '@angular/material/slide-toggle';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatSort } from '@angular/material/sort';
+import { MatTableDataSource } from '@angular/material/table';
 import {DbService} from '../../core/db.service';
 import {UserData} from '../../entities/user';
 import {NgxSpinnerService} from 'ngx-spinner';
@@ -14,8 +18,8 @@ export class UsersTableComponent implements AfterViewInit {
   displayedColumns: string[] = ['name', 'faculty', 'email', 'admin'];
   dataSource: MatTableDataSource<UserData>;
 
-  @ViewChild(MatPaginator) paginator: MatPaginator;
-  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatPaginator, {static: true}) paginator: MatPaginator;
+  @ViewChild(MatSort, {static: true}) sort: MatSort;
 
   constructor(private db: DbService, private spinner: NgxSpinnerService, private snackBar: MatSnackBar) {}
 
