@@ -3,7 +3,7 @@ import {DbService} from '../../core/db.service';
 import {PdfService} from '../pdf.service';
 import {CourseWithFaculty} from '../../entities/course';
 import {UploadService} from '../upload.service';
-import {MatSnackBar} from '@angular/material';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import {ActivatedRoute} from '@angular/router';
 import {OCRService} from '../../core/ocr.service';
 import {NgxSpinnerService} from 'ngx-spinner';
@@ -47,9 +47,9 @@ enum UploadState {
 })
 export class UploadComponent implements OnInit {
 
-  @ViewChild('file') file;
-  @ViewChild('imagesCollpaseTrigger') imagesCollpaseTrigger;
-  @ViewChild('collapseOne') collapseOneTrigger;
+  @ViewChild('file', {static: true}) file;
+  @ViewChild('imagesCollpaseTrigger', {static: true}) imagesCollpaseTrigger;
+  @ViewChild('collapseOne', {static: true}) collapseOneTrigger;
 
   public questions: QuestionSolution[] = [];
   private activeQuestion = 0;
