@@ -15,7 +15,11 @@ import {QuestionComponent} from './ui/question/question.component';
 import {ExamComponent} from './ui/exam/exam.component';
 import {HeaderComponent} from './ui/header/header.component';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {MatButtonModule, MatIconModule, MatInputModule, MatMenuModule, MatPaginatorModule, MatTableModule} from '@angular/material';
+import {MatPaginatorModule, MatTableModule} from '@angular/material';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
+import { MatMenuModule } from '@angular/material/menu';
 import {MatCheckboxModule} from '@angular/material/checkbox';
 import {CoreModule} from './core/core.module';
 import {FormsModule} from '@angular/forms';
@@ -29,7 +33,7 @@ import {UploadModule} from './upload/upload.module';
 import {AdminModule} from './admin/admin.module';
 import {DiscussionsModule} from './discussions/discussions.module';
 import {GamificationModule} from './gamification/gamification.module';
-import {MatSortModule} from '@angular/material';
+import { MatSortModule } from '@angular/material/sort';
 import {PendingScansModule} from './pending-scans/pending-scans.module';
 import {NgbModule, NgbRatingModule} from '@ng-bootstrap/ng-bootstrap';
 import {NgbModalModule} from '@ng-bootstrap/ng-bootstrap';
